fix function position
