+++ src/client/convex/etc.ts
@@ -1,5 +1,10 @@
 import * as Discord from 'discord.js'
+import Option from 'type-of-option'
 import Settings from 'const-settings'
+import {AtoA} from 'alphabet-to-number'
+import * as status from '../../io/status'
+import * as spreadsheet from '../../util/spreadsheet'
+import {User} from '../../io/type'
 import * as util from '../../util'
 
 /**
@@ -46,4 +51,96 @@ export const AddTaskKillRoll = async (msg: Discord.Message) => {
 
     msg.reply('タスキルロールを付けたわよ！')
   }
+}
+
+/**
+ * 凸残ロールを全て外す
+ * @param msg DiscordからのMessage
+ */
+export const RemoveRole = async (msg: Discord.Message) => {
+  // べろばあのクランメンバー一覧を取得
+  const clanMembers =
+    util
+      .GetGuild()
+      ?.roles.cache.get(Settings.ROLE_ID.CLAN_MEMBERS)
+      ?.members.map(m => m) ?? []
+
+  // クランメンバーの凸残ロールを全て外す
+  await Promise.all(clanMembers.map(async m => await m?.roles.remove(Settings.ROLE_ID.REMAIN_CONVEX)))
+
+  msg.reply('凸残ロール全て外したわよ！')
+}
+
+/**
+ * スプレッドシートのメンバー一覧を更新する
+ * @param msg DiscordからのMessage
+ */
+export const UpdateMembers = async (msg: Discord.Message) => {
+  // クランメンバー一覧をニックネームで取得
+  const users: Option<User[]> = msg.guild?.roles.cache
+    .get(Settings.ROLE_ID.CLAN_MEMBERS)
+    ?.members.map(m => ({
+      name: util.GetUserName(m),
+      id: m.id,
+      limit: '',
+    }))
+    .sort((a, b) => (a.name > b.name ? 1 : -1)) // 名前順にソート
+
+  // ステータスを更新
+  await status.UpdateUsers(users)
+  await util.Sleep(100)
+
+  // スプレッドシートに名前とidを保存する
+  await fetchNameAndID(users, Settings.INFORMATION_SHEET.SHEET_NAME)
+
+  msg.reply('クランメンバー一覧を更新したわよ！')
+}
+
+/**
+ * 妹クランのメンバー一覧を更新する
+ * @param msg DiscordからのMessage
+ */
+export const UpdateSisters = async (msg: Discord.Message) => {
+  // 妹クランメンバー一覧をニックネームで取得
+  const users: Option<User[]> = msg.guild?.roles.cache
+    .get(Settings.ROLE_ID.SISTER_MEMBERS)
+    ?.members.map(m => ({
+      name: util.GetUserName(m),
+      id: m.id,
+      limit: '',
+    }))
+    .sort((a, b) => (a.name > b.name ? 1 : -1)) // 名前順にソート
+
+  // スプレッドシートに名前とidを保存する
+  await fetchNameAndID(users, Settings.SISTER_SHEET.SHEET_NAME)
+
+  msg.reply('妹クランメンバー一覧を更新したわよ！')
+}
+
+/**
+ * 指定されたシートにメンバーの名前とidを保存する
+ * @param members メンバーの情報
+ * @param name 書き込むシートの名前
+ */
+const fetchNameAndID = async (users: Option<User[]>, name: string) => {
+  // 値がない場合は終了
+  if (!users) return
+
+  // 書き込み先のシートを取得
+  const sheet = await spreadsheet.GetWorksheet(name)
+
+  // メンバー一覧を更新
+  await Promise.all(
+    users.map(async (m, i) => {
+      const col = Settings.INFORMATION_SHEET.MEMBER_COLUMN
+
+      // 名前を更新
+      const name_cell = await sheet.getCell(`${col}${i + 3}`)
+      name_cell.setValue(m.name)
+
+      // idを更新
+      const id_cell = await sheet.getCell(`${AtoA(col, 1)}${i + 3}`)
+      id_cell.setValue(m.id)
+    })
+  )
 }
